refactor: headerをtsに移行
